+++ src/Collections.py
@@ -1,3 +1,5 @@
+from typing import List
+from PySide6.QtCore import Slot
 from PySide6.QtWidgets import QListWidget, QListWidgetItem, QStyleOptionTab, QVBoxLayout, QLabel, QListView, QWidget
 from collections import defaultdict # might not need this
 from functools import lru_cache
@@ -31,16 +33,20 @@ class Collection(QListWidgetItem):
 
 
 class Collections(QListWidget): # Displays collections
-    def __init__(self, songs_ref: Songs):
+    def __init__(self):
         super().__init__()
         self.setSpacing(5)
         self.setWrapping(True)
+        self.songs_ref: Songs
 
         self.label = QLabel("Collections")
         self.setVisible(True)
 
         self.populate(self.load_collections())
-        self.itemClicked.connect(popu)
+        self.itemClicked.connect(self.populate_from_collection)
+
+    def set_songs_ref(self, songs: Songs):
+        self.songs_ref = songs
 
     def populate(self, collection_list: list['Collection']):
         self.clear()
@@ -48,12 +54,16 @@ class Collections(QListWidget): # Displays collections
             item.setText(item.name)
             self.insertItem(row, item)
 
+    def populate_from_collection(self, item):
+        self.songs_ref.populate(self.get_songs_from_collection(item))
+
+
     def get_songs_from_collection(self, item: Collection) -> list['Song']: # This could be just id?
         #Get ID when item clicked
         #Get all songs with that Song-Collection Relationship
         #Create song object for each and populate collections with that
 
-        song_list = [Song()]
+        song_list = []
         with SongDB() as db:
             id_list = Songs_Collections().read(0, [item.id]).fetchall()
             id_list = [x[0] for x in id_list]
@@ -77,3 +87,21 @@ class Collections(QListWidget): # Displays collections
                 # add to list
                 load.append(inst)
         return load
+
+class CollectionsPane(QWidget):
+    def __init__(self):
+        super().__init__()
+        layout = QVBoxLayout()
+
+        # Get collections
+        self.collections = Collections()
+        # Set label for widget
+        label = QLabel()
+        label.setText("Collections")
+
+        # Add label and widget to layout
+        layout.addWidget(label)
+        layout.addWidget(self.collections)
+
+        # apply layout to instance of Collections pane (self)
+        self.setLayout(layout)
